% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DocumentationExamplesQiuDocument.R
\docType{data}
\name{example74c}
\alias{example74c}
\title{Data from Example 7.4(c) Qiu (2014).}
\format{
The data (X1,X2,X3) consist of 30 observations each variable.

\describe{
 \item{X}{1st batch.}
 \item{X.1}{2nd batch.}
 \item{X.2}{3rd batch.}
}
}
\source{
\url{https://users.phhp.ufl.edu/pqiu/research/book/spc/data/example74c.dat}
}
\usage{
example74c
}
\description{
A dataset containing the data set used in Example 7.4(c) of Qiu (2014).
}
\keyword{datasets}
