% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DocumentationExamplesQiuDocument.R
\docType{data}
\name{example81}
\alias{example81}
\title{Data from Example 8.1 on page 319 Qiu (2014).}
\format{
A data frame with 300 rows (30 batches of size equals to 10)
\describe{
 \item{X}{observations of all batches}
 \item{X.id}{id of each observation of the batch}
}
}
\source{
\url{https://users.phhp.ufl.edu/pqiu/research/book/spc/data/example81.dat}
}
\usage{
example81
}
\description{
A dataset containing the data set used in Example 8.1 on page 319 of Qiu (2014).
}
\keyword{datasets}
