% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DocumentationExamplesQiuDocument.R
\docType{data}
\name{example82}
\alias{example82}
\title{Data from Example 8.2 Qiu (2014).}
\format{
A data frame with 150 rows (30 batches of size equals to 5)
\describe{
 \item{X}{observations of all batches}
 \item{X.id}{id of each observation of the batch}
}
}
\source{
\url{https://users.phhp.ufl.edu/pqiu/research/book/spc/data/example82.dat}
}
\usage{
example82
}
\description{
A dataset containing the data set used in Example 8.2 on page 323 of Qiu (2014).
}
\keyword{datasets}
