% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DocumentationExamplesQiuDocument.R
\docType{data}
\name{example87}
\alias{example87}
\title{Data from Example 8.7 on page 339 Qiu (2014).}
\format{
A data frame with 86 rows (86 batches of size equals to 1)
\describe{
 \item{X}{observations of all batches}
 \item{X.id}{id of each observation of the batch}
 \item{Y}{reference sample of size equals to 14}
}
}
\source{
\url{https://users.phhp.ufl.edu/pqiu/research/book/spc/data/example87.dat}
}
\usage{
example87
}
\description{
A dataset containing the data set used in Example 8.7 on page 339 of Qiu (2014).
}
\keyword{datasets}
