% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DocumentationExamplesQiuDocument.R
\docType{data}
\name{example91}
\alias{example91}
\title{Data from Example 9.1 on page 369 Qiu (2014).}
\format{
The data (X,Y) consist of 20 batches with 50 observations in each batch.

\describe{
 \item{V1}{1st batch.}
 \item{V2}{2nd batch.}
}
}
\source{
\url{https://users.phhp.ufl.edu/pqiu/research/book/spc/data/example91.dat}
}
\usage{
example91
}
\description{
A dataset containing the data set used in Example 9.1 on page 369 of Qiu (2014).
}
\keyword{datasets}
