% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DocumentationExamplesQiuDocument.R
\docType{data}
\name{example93}
\alias{example93}
\title{Data from Example 9.3 Qiu (2014).}
\format{
The data (X,Y) consist of 20 batches with 10 observations in each batch.

\describe{
 \item{X}{1st batch.}
 \item{X.1}{2nd batch.}
}
}
\source{
\url{https://users.phhp.ufl.edu/pqiu/research/book/spc/data/example93.dat}
}
\usage{
example93
}
\description{
A dataset containing the data set used in Example 9.3 of Qiu (2014).
}
\keyword{datasets}
