% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UnivariateNormalScores.R
\name{srank}
\alias{srank}
\title{Sequential Rank}
\usage{
srank(X, Y = NULL)
}
\arguments{
\item{X}{vector. New observations to obtain the N¡normal scores.}

\item{Y}{vector. If \code{Y} is not defined (no previous observation available, \code{NULL}), NS is relative to \code{X}. Default \code{NULL}.}
}
\value{
vector. Sequentil Ranks for the \code{X} observations. If ties occurs, average of the ranks are used.
}
\description{
Get the sequential rank of observations (\code{X})
relative to previous observations (\code{Y}).
}
\examples{
X <- c(30, 35, 45)
srank(X)
}
